Parsing orbit vectors using longs
Using ints was causing overflow, and a pretty bad error.

I am now catching the overflow error, and surviving it.
After all, we are only parsing automatic orbits that may
not even be needed to interpret the history.

+++ core/src/main/java/com/vzome/core/algebra/PlasticPhiField.java
@@ -151,9 +151,9 @@ public class PlasticPhiField extends ParameterizedField
     }
     
     @Override
-    protected int[] convertGoldenNumberPairs(int[] terms) {
+    protected long[] convertGoldenNumberPairs(long[] terms) {
         if (terms.length == 2) {
-            terms = new int[] {
+            terms = new long[] {
                     terms[0],           // units
                     terms[1],           // phis
                     0,   // zero fill the rest...
